eber-tether as peerDep and tooltip-on-parent added
