+++ app/components/tooltip-on-parent.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-tooltips/components/tooltip-on-parent';
