+++ package.json
@@ -39,7 +39,6 @@
     "ember-load-initializers": "^0.5.1",
     "ember-paper": "0.2.14",
     "ember-resolver": "^2.0.3",
-    "ember-tether": "0.3.1",
     "ember-try": "^0.2.2",
     "loader.js": "^4.0.1"
   },
@@ -54,6 +53,9 @@
   "dependencies": {
     "ember-cli-babel": "^5.1.6",
     "ember-cli-sass": "5.3.1"
+  },
+  "peerDependencies": {
+    "ember-tether": "0.3.1"
   },
   "ember-addon": {
     "configPath": "tests/dummy/config",
