+++ tests/dummy/app/templates/index.hbs
@@ -1,8 +1,8 @@
 {{#test-component}}
   Sup
-  {{#tooltip-on-element side='left'}}
+  {{#tooltip-on-parent side='left'}}
     Hey sdg sdg
-  {{/tooltip-on-element}}
+  {{/tooltip-on-parent}}
 {{/test-component}}
 
 <br>
